Renamed code_parse file to be more generic
